# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TJ-T-Mart API",
        "description": "REST API for TJ-T-Mart — Dormitory Mini-Mart System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/token": {
            "get": {
                "tags": [
                    "Token"
                ],
                "summary": "List user's token transactions",
                "operationId": "470e4f3a045ba1884a96a0754f403e22",
                "responses": {
                    "200": {
                        "description": "Paginated token history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Token"
                ],
                "summary": "Purchase electricity token",
                "description": "Generates a 20-digit token instantly. Status set to completed.",
                "operationId": "fd9d84f0cd1f43d7f185f1eea43e330d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TokenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Token purchased",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ApiSuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "TokenRequest": {
                "required": [
                    "nominal",
                    "metode_pembayaran"
                ],
                "properties": {
                    "nominal": {
                        "type": "integer",
                        "example": 50000,
                        "enum": [
                            20000,
                            50000,
                            100000,
                            200000,
                            500000
                        ]
                    },
                    "metode_pembayaran": {
                        "type": "string",
                        "example": "COD",
                        "enum": [
                            "COD",
                            "MIDTRANS"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}